Add test_sort.py and separate test for bubble sort
