+++ sort/bubble_sort.py
@@ -21,13 +21,3 @@ def bubble_sort(arr):
                 swap(i - 1, i)
                 swapped = True
     return arr
-
-class TestSuite(unittest.TestCase):
-    """
-        test suite for the function (above)
-    """
-    def test_bubble_sort(self):
-        self.assertEqual([1, 5, 23, 57, 65, 1232],
-                         bubble_sort([1, 5, 65, 23, 57, 1232]))
-if __name__ == "__main__":
-    unittest.main()
